requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.1
functions-framework==2.2.1
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-bigquery==2.22.0
google-cloud-core==1.7.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
httpagentparser==1.9.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
watchdog==1.0.2
Werkzeug==2.0.1
